+++ crates/config/src/lib.rs
@@ -1,22 +1,22 @@
 mod errors;
-mod task;
+mod package;
+mod project;
 mod types;
 mod validators;
+mod workspace;
 
 pub mod constants;
-pub mod global_project;
-pub mod package;
-pub mod project;
-pub mod workspace;
 
-// Re-exports structs for convenience
-pub use global_project::GlobalProjectConfig;
 pub use package::{PackageJson, PackageJsonValue};
-pub use project::{ProjectConfig, ProjectType};
-pub use task::{TaskConfig, TaskMergeStrategy, TaskOptionsConfig, TaskType};
+pub use project::global::GlobalProjectConfig;
+pub use project::task::{TaskConfig, TaskMergeStrategy, TaskOptionsConfig, TaskType};
+pub use project::{ProjectConfig, ProjectMetadataConfig, ProjectType};
 pub use types::{FileGroups, FilePath, FilePathOrGlob, ProjectID, TargetID, TaskID};
 pub use validator::ValidationErrors;
-pub use workspace::{VcsConfig, VcsManager, WorkspaceConfig};
+pub use workspace::{
+    NodeConfig, NpmConfig, PackageManager, PnpmConfig, VcsConfig, VcsManager, WorkspaceConfig,
+    YarnConfig,
+};
 
 pub fn load_workspace_config_template() -> &'static str {
     include_str!("../templates/workspace.yml")
